# Remove @Singleton annotations.