… src/main/protobuf/jars/libkotlin_model_proto-speed.jar