This is a change that removes all singleton annotations. I thought that this could actually effect performance, but it appears there is no change (see below). But what removing singletons does is reduce the likelihood of "state bleed-over" during compilations.

I ran a series of timing tests with and without singleton annotations. ``bazel clean --expunge`` was executed between each run. Command used was ``time make``.

With Singleton annotations:
```
make  0.05s user 0.05s system 0% cpu 2:54.62 total
make  0.05s user 0.05s system 0% cpu 2:39.98 total
make  0.05s user 0.05s system 0% cpu 2:42.00 total
```
*Without* Singleton annotations:
```
make  0.05s user 0.05s system 0% cpu 2:46.34 total
make  0.05s user 0.05s system 0% cpu 2:41.53 total
make  0.05s user 0.05s system 0% cpu 2:41.74 total
```
As can be seen, there is little to no difference between the two, except the use without Singletons is likely safer.