@brettwooldridge 2022-04-05
@cgruber As long as I have your ear, is there any documentation regarding how the runtime structure of classes is assembled?

In particular, I am trying to understand how persistent workers actually work. In Buck, we are able to run the K2JVMCompiler "in-process" with Buck, i.e. executing within the same JVM instance that is running Buck itself. However, in Bazel, when I run ``ps`` during compilation, it appears that each "worker" is forked as a separate JVM process:
```
brettw            3788   0.0  2.2 26245504 1452628   ??  S     5:08AM   8:38.28 /private/var/tmp/_bazel_brettw/ba6f1f31eab91f1d52ff2982674ba73f/execroot/__main__/bazel-out/host/bin/external/io_bazel_rules_kotlin/src/main/kotlin/build.runfiles/remotejdk11_macos/bin/java -classpath bazel-out/host/bin/external/io_bazel_rules_kotlin/src/main/kotlin/build.runfiles/__main__/../io_bazel_rules_kotlin/src/main/kotlin/kotlin_worker.jar:bazel-out/host/bin/external/io_bazel_rules_kotlin/src/main/kotlin/build.runfiles/__main__/../remote_java_tools/java_tools/JacocoCoverage_jarjar_deploy.jar -XX:ErrorFile=/dev/stderr io.bazel.kotlin.builder.cmd.Build --persistent_worker
```
If this is so, how is "work" passed to each worker process? We are suspecting that the JIT-compilation of the K2JVMCompiler that must occur within each independent JVM instance is substantially slowing down the build in Bazel compared to our Buck build. Under Buck we were seeing 1-minute total clean build times, compared to 3-minute total clean build times under Bazel.