@rockwotj 2022-04-06
@brettwooldridge see these docs: https://docs.bazel.build/versions/main/persistent-workers.html

Essential Protos are passed over stdin/stdout