@restingbull 2022-04-07
> @cgruber As long as I have your ear, is there any documentation regarding how the runtime structure of classes is assembled?
> 
> In particular, I am trying to understand how persistent workers actually work. In Buck, we are able to run the K2JVMCompiler "in-process" with Buck, i.e. executing within the same JVM instance that is running Buck itself. However, in Bazel, when I run `ps` during compilation, it appears that each "worker" is forked as a separate JVM process:
> 
> ```
> brettw           21301 142.4  1.3 26162484 890544   ??  S     5:04PM   0:48.33 /private/var/tmp/_bazel_brettw/2eb6368d09513d6b04f237307bef6bb8/execroot/dev_io_bazel_rules_kotlin/bazel-out/host/bin/src/main/kotlin/io/bazel/kotlin/builder/cmd/build.runfiles/remotejdk11_macos/bin/java -classpath bazel-out/host/bin/src/main/kotlin/io/bazel/kotlin/builder/cmd/build.runfiles/dev_io_bazel_rules_kotlin/src/main/kotlin/io/bazel/kotlin/builder/cmd/build_jarjar.jar -XX:ErrorFile=/dev/stderr -XX:+IgnoreUnrecognizedVMOptions --add-opens=java.base/java.nio=ALL-UNNAMED --add-opens=java.base/java.lang=ALL-UNNAMED --add-opens=jdk.jdeps/com.sun.tools.jdeps=ALL-UNNAMED io.bazel.kotlin.builder.cmd.Build --persistent_worker
> ```
> 
> If this is so, how is "work" passed to each worker process? We are suspecting that the JIT-compilation of the K2JVMCompiler that must occur within each independent JVM instance is substantially slowing down the build in Bazel compared to our Buck build. Under Buck we were seeing 1-minute total clean build times, compared to 3-minute total clean build times under Bazel.

I'm currently working on removing the indirection. The initial setup was arranged to allow swapping the compiler implementation via classpaths. This is no longer necessary.